+++ bench_results/source_doublet_single_oct.cpp
@@ -12,14 +12,14 @@
 #include <Eigen/Core>
 #include <algorithm>
 
-Eigen::MatrixXd SourceDoubletSingleOct::assembleLhs() {
+Eigen::MatrixXf SourceDoubletSingleOct::assembleLhs() {
 
   std::size_t evalDims = evalPointsRef.get().mEvalPoints.rows();
   std::size_t wakeDims = wakePanelCompTasks.size();
-  Eigen::MatrixXd surfaceInfluenceMatrix = makeInfluenceMatrix<DoubletP, true>(
+  Eigen::MatrixXf surfaceInfluenceMatrix = makeInfluenceMatrix<DoubletP, true>(
       evalDims, evalDims, surfacePanelCompTasks);
   std::cout << surfaceInfluenceMatrix.topLeftCorner(10, 10) << "\n";
-  Eigen::MatrixXd wakeInfluenceMatrix = makeInfluenceMatrix<DoubletP, false>(
+  Eigen::MatrixXf wakeInfluenceMatrix = makeInfluenceMatrix<DoubletP, false>(
       evalDims, wakeDims, wakePanelCompTasks);
 
   // combine source and wake
@@ -42,16 +42,16 @@ Eigen::MatrixXd SourceDoubletSingleOct::assembleLhs() {
   return surfaceInfluenceMatrix;
 }
 
-Eigen::VectorXd SourceDoubletSingleOct::assembleRhs() {
+Eigen::VectorXf SourceDoubletSingleOct::assembleRhs() {
   std::size_t evalDims = evalPointsRef.get().mEvalPoints.rows();
-  Eigen::MatrixXd sourceInfluenceMat = makeInfluenceMatrix<SourceP, true>(
+  Eigen::MatrixXf sourceInfluenceMat = makeInfluenceMatrix<SourceP, true>(
       evalDims, evalDims, surfacePanelCompTasks);
   sourceStrength = rowwiseDotProduct(IPM::surfacePanelRef.get().normalVectors,
                                      IPM::freeStream);
   return -(sourceInfluenceMat * sourceStrength);
 }
 
-Eigen::MatrixXd SourceDoubletSingleOct::calculatePanelVelocities() {
+Eigen::MatrixXf SourceDoubletSingleOct::calculatePanelVelocities() {
   auto &panel = IPM::surfacePanelRef.get();
   std::size_t nYSecs = panel.mSurface.nYsecs;
   std::size_t nXsecs = panel.mSurface.nXsecs;
@@ -64,7 +64,7 @@ Eigen::MatrixXd SourceDoubletSingleOct::calculatePanelVelocities() {
   // order, yes compute task <- centerpoints <- face
 
   // 1): Convert the centrepoints to panel reference frame
-  Eigen::ArrayXXd xPoints(nXsecs, nYSecs);
+  Eigen::ArrayXXf xPoints(nXsecs, nYSecs);
   xPoints.setZero();
   for (int iY = 0; iY < nYSecs; iY++) {
     for (int iX = 1; iX < nXsecs; iX++) {
@@ -75,7 +75,7 @@ Eigen::MatrixXd SourceDoubletSingleOct::calculatePanelVelocities() {
                         xPoints(iX - 1, iY);
     }
   }
-  Eigen::ArrayXXd yPoints(nXsecs, nYSecs);
+  Eigen::ArrayXXf yPoints(nXsecs, nYSecs);
   yPoints.setZero();
   for (int iX = 0; iX < nXsecs; iX++) {
     for (int iY = 1; iY < nYSecs; iY++) {
@@ -88,14 +88,14 @@ Eigen::MatrixXd SourceDoubletSingleOct::calculatePanelVelocities() {
   }
 
   // 2:) u = -d(mu)/d(x_l); v = -d(mu)/d(y_l); w = sigma
-  Eigen::ArrayXXd fPoints(nXsecs, nYSecs);
+  Eigen::ArrayXXf fPoints(nXsecs, nYSecs);
   fPoints << IPM::solution.reshaped(nXsecs, nYSecs); // no minus
 
-  Eigen::ArrayX3d inducedVelocities(nXsecs * nYSecs, 3);
+  Eigen::ArrayX3f inducedVelocities(nXsecs * nYSecs, 3);
   inducedVelocities << -centralDifference<true>(xPoints, fPoints).reshaped(),
       -centralDifference<false>(yPoints, fPoints).reshaped(), -sourceStrength;
 
-  Eigen::ArrayX3d globalVelocites(nXsecs * nYSecs, 3);
+  Eigen::ArrayX3f globalVelocites(nXsecs * nYSecs, 3);
 
   globalVelocites << rowwiseDotProduct(panel.tangentXVectors, IPM::freeStream),
       rowwiseDotProduct(panel.tangentYVectors, IPM::freeStream),
@@ -123,8 +123,8 @@ Eigen::MatrixXd SourceDoubletSingleOct::calculatePanelVelocities() {
 SourceDoubletSingleOct::SourceDoubletSingleOct(
     const PanelGeometry<SurfacePanel> &surfacePanelGeo,
     const PanelGeometry<WakePanel> &wakePanelGeo,
-    const EvalPoints<double> &evalPoints, std::unique_ptr<ISolver> &&solver,
-    double AoAd)
+    const EvalPoints<float> &evalPoints, std::unique_ptr<ISolver> &&solver,
+    float AoAd)
     : IPanelMethod(surfacePanelGeo, wakePanelGeo, evalPoints,
                    std::move(solver)) {
   IPM::setFlowParams(AoAd);
@@ -133,11 +133,11 @@ SourceDoubletSingleOct::SourceDoubletSingleOct(
 template <typename PanelGeo>
 std::vector<std::vector<std::size_t>>
 getAllNeighbours(const PanelGeo &surfacePanelGeo,
-                 const EvalPoints<double> &evalPoints,
-                 const Eigen::ArrayXd &searchRadius) {
+                 const EvalPoints<float> &evalPoints,
+                 const Eigen::ArrayXf &searchRadius) {
 
   const auto cloud = convertMat2Cloud(evalPoints.mEvalPoints);
-  double resolution = 100;
+  float resolution = 100;
 
   Octree octree(resolution);
   octree.setInputCloud(cloud);
@@ -149,7 +149,7 @@ getAllNeighbours(const PanelGeo &surfacePanelGeo,
 }
 
 template <typename PanelGeo>
-auto getPanelDiagonalLength(const PanelGeo &panelGeo) -> Eigen::ArrayXd {
+auto getPanelDiagonalLength(const PanelGeo &panelGeo) -> Eigen::ArrayXf {
   const auto &faceIdx = panelGeo.mSurface.mFaceNodeIdx;
   const auto &surfPoints = panelGeo.mSurface.mPoints;
 
@@ -161,7 +161,7 @@ auto getPanelDiagonalLength(const PanelGeo &panelGeo) -> Eigen::ArrayXd {
 
 template <typename PanelGeo>
 auto makeComputeTasks(const PanelGeo &panelGeo,
-                      const EvalPoints<double> &evalPoints) {
+                      const EvalPoints<float> &evalPoints) {
   int nPanels = panelGeo.centrePoints.rows();
   auto faceIdxs = RANGE(nPanels);
   auto allNeighbours = getAllNeighbours(panelGeo, evalPoints,
@@ -175,8 +175,8 @@ auto makeComputeTasks(const PanelGeo &panelGeo,
   print("Full Search Space: ",
         evalPoints.mEvalPoints.rows() * panelGeo.centrePoints.rows());
   print("Potential Speed Up: ",
-        (double)(evalPoints.mEvalPoints.rows() * panelGeo.centrePoints.rows()) /
-            (double)totalNum);
+        (float)(evalPoints.mEvalPoints.rows() * panelGeo.centrePoints.rows()) /
+            (float)totalNum);
   return initialize_transform<std::vector<ComputeTask>>(
       faceIdxs.begin(), faceIdxs.end(), [&](int faceIdx) {
         return createInfluenceComputeTask(panelGeo, evalPoints, faceIdx,
@@ -196,6 +196,6 @@ void SourceDoubletSingleOct::run() {
   IPM::velocities = calculatePanelVelocities();
 }
 
-Eigen::ArrayXd SourceDoubletSingleOct::getSource() const {
+Eigen::ArrayXf SourceDoubletSingleOct::getSource() const {
   return sourceStrength;
 }
